+++ src/NServiceBus.AcceptanceTesting/Support/EndpointRunner.cs
@@ -46,7 +46,6 @@
                     throw new Exception($"Missing EndpointSetup<T> in the constructor of {endpointName} endpoint.");
                 }
                 var endpointConfiguration = await configuration.GetConfiguration(run).ConfigureAwait(false);
-                RegisterInheritanceHierarchyOfContextInSettings(scenarioContext, endpointConfiguration);
                 TrackFailingMessages(endpointName, endpointConfiguration);
 
                 if (!string.IsNullOrEmpty(configuration.CustomMachineName))
@@ -55,6 +54,8 @@
                 }
 
                 endpointConfiguration.EnableFeature<FeatureStartupTaskRunner>();
+                endpointConfiguration.GetSettings().Set(scenarioContext); // make scenario context available to the ScenarioContextRegistration feature
+                endpointConfiguration.EnableFeature<ScenarioContextRegistration>();
 
                 endpointBehavior.CustomConfig.ForEach(customAction => customAction(endpointConfiguration, scenarioContext));
 
@@ -93,16 +94,6 @@
                 return Task.FromResult(0);
             }));
             endpointConfiguration.Pipeline.Register(new CaptureExceptionBehavior(scenarioContext.UnfinishedFailedMessages), "Captures unhandled exceptions from processed messages for the AcceptanceTesting Framework");
-        }
-
-        void RegisterInheritanceHierarchyOfContextInSettings(ScenarioContext context, EndpointConfiguration endpointConfiguration)
-        {
-            var type = context.GetType();
-            while (type != typeof(object))
-            {
-                endpointConfiguration.GetSettings().Set(type.FullName, scenarioContext);
-                type = type.BaseType;
-            }
         }
 
         public override async Task Start(CancellationToken cancellationToken = default)
